develop <- feature/quoteTable

+++ springboot/src/main/java/ca/jrvs/apps/trading/model/IexQuote.java
@@ -116,11 +116,11 @@ public class IexQuote {
   @JsonProperty("avgTotalVolume")
   private Long avgTotalVolume;
   @JsonProperty("iexBidPrice")
-  private Long iexBidPrice;
+  private Double iexBidPrice;
   @JsonProperty("iexBidSize")
   private Long iexBidSize;
   @JsonProperty("iexAskPrice")
-  private Long iexAskPrice;
+  private Double iexAskPrice;
   @JsonProperty("iexAskSize")
   private Long iexAskSize;
   @JsonProperty("marketCap")
@@ -420,9 +420,7 @@ public class IexQuote {
   }
 
   @JsonProperty("iexMarketPercent")
-  public void setIexMarketPercent(BigDecimal iexMarketPercent) {
-    this.iexMarketPercent = iexMarketPercent;
-  }
+  public void setIexMarketPercent(BigDecimal iexMarketPercent) { this.iexMarketPercent = iexMarketPercent; }
 
   @JsonProperty("iexVolume")
   public Long getIexVolume() {
@@ -445,14 +443,12 @@ public class IexQuote {
   }
 
   @JsonProperty("iexBidPrice")
-  public Long getIexBidPrice() {
+  public Double getIexBidPrice() {
     return iexBidPrice;
   }
 
   @JsonProperty("iexBidPrice")
-  public void setIexBidPrice(Long iexBidPrice) {
-    this.iexBidPrice = iexBidPrice;
-  }
+  public void setIexBidPrice(Double iexBidPrice) { this.iexBidPrice = iexBidPrice; }
 
   @JsonProperty("iexBidSize")
   public Long getIexBidSize() {
@@ -465,12 +461,10 @@ public class IexQuote {
   }
 
   @JsonProperty("iexAskPrice")
-  public Long getIexAskPrice() {
-    return iexAskPrice;
-  }
+  public Double getIexAskPrice() { return iexAskPrice; }
 
   @JsonProperty("iexAskPrice")
-  public void setIexAskPrice(Long iexAskPrice) {
+  public void setIexAskPrice(Double iexAskPrice) {
     this.iexAskPrice = iexAskPrice;
   }
